moves outputdispatcher to the server module
