--- client/src/main/java/connection/OutputSelector.java
@@ -1,7 +1,0 @@
-package connection;
-
-public interface OutputSelector {
-
-  void printToDesiredOutput(String message);
-
-}
