--- client/src/main/java/connection/SysOut.java
@@ -1,19 +1,0 @@
-package connection;
-
-import connection.OutputChannelDispatcher;
-import connection.OutputSelector;
-
-public class SysOut implements OutputSelector {
-
-  OutputChannelDispatcher outputChannelDispatcher;
-
-  public SysOut(OutputChannelDispatcher outputChannelDispatcher) {
-    this.outputChannelDispatcher = outputChannelDispatcher;
-  }
-
-  @Override
-  public void printToDesiredOutput(String message) {
-    System.out.println(message);
-
-  }
-}
