+++ client/src/main/java/gui/building/FleetDropController.java
@@ -2,7 +2,6 @@ package gui.building;
 
 import connection.Client;
 import connection.FleetSender;
-import connection.OutputChannelDispatcher;
 import gui.fields.BoundField;
 import gui.fields.FieldSize;
 import gui.fields.Mast;
@@ -45,7 +44,6 @@ public class FleetDropController implements Initializable {
   private Fleet fleet;
   private final Client client;
   private LanguageVersion languageVersion = new LanguageVersion();
-  private OutputChannelDispatcher outputChannelDispatcher = new OutputChannelDispatcher();
 
   @FXML
   private Button connectButton;
@@ -177,7 +175,6 @@ public class FleetDropController implements Initializable {
           nextButton.fireEvent(new ConnectEvent());
         } else {
           info.setText(languageVersion.getMessage("emptyFleetInfo"));
-          outputChannelDispatcher.printToDesiredOutput(languageVersion.getMessage("emptyFleetInfo"));
         }
         event.consume();
       };
