+++ server/src/main/java/communication/OutputSelector.java
@@ -1,0 +1,7 @@
+package communication;
+
+public interface OutputSelector {
+
+  void writeMessage(String message);
+
+}
