+++ server/src/main/java/communication/SysOut.java
@@ -1,0 +1,10 @@
+package communication;
+
+public class SysOut implements OutputSelector {
+
+  @Override
+  public void writeMessage(String message) {
+    System.out.println(message);
+
+  }
+}
