+++ server/src/main/java/communication/SystemErr.java
@@ -1,0 +1,9 @@
+package communication;
+
+public class SystemErr implements OutputSelector {
+
+  @Override
+  public void writeMessage(String message) {
+    System.err.println(message);
+  }
+}
